record: allow unspecified values
In other languages it's possible to ignore empty/null JSON fields when
building the resulting JSON object. For example, in Go you would specify
`omitempty` in the struct field's tag. In Java, you would use
"@JsonInclude(Include.NON_NULL)" on a field or class to indicate that null
fields should not be serialized.

So, when using records as object with guile-json you can no set a field
to *unspecified*. When that happens the field won't be serialzied to JSON.

* json/record.scm (define-json-writer): ignore fields with *unspecified*
values.

* json/record.scm (define-json-reader): assign *unspecified* if field not
found in object's alist.